updated discovery code and added comments
